Ontology_Factory/preprocess/requirements.txt
pyyaml>=6.0.1
httpx>=0.27.0
pypdf>=4.2.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
charset-normalizer>=3.3.2
